Simple library and app and Makefile
For now, a one-class library called "TestLib", which just instances
a std::string and std::vector<int>, with simple set/get and
populate/show methods.

Also, a simple app "testapp.cxx" to exercise the library class.

A simple Makefile that builds the lib, and separate target for the app
so the app can be built with other compilers.

Currently limited to:
	> Unix builds
	> Building a static lib (.a)

TODO:
	Add compiler version# recording during lib build
	Add compiler version# checking during app build